Update document space to latest api revision

+++ src/pages/DocumentSpace/DocumentSpacePage.tsx
@@ -1,6 +1,6 @@
 import { none, useHookstate } from '@hookstate/core';
 import { IDatasource, ValueFormatterParams } from 'ag-grid-community';
-import React, { ChangeEvent, useEffect } from 'react';
+import React, { ChangeEvent, useEffect, useRef } from 'react';
 import Button from '../../components/Button/Button';
 import { InfiniteScrollOptions } from '../../components/DataCrudFormPage/infinite-scroll-options';
 import DeleteCellRenderer from '../../components/DeleteCellRenderer/DeleteCellRenderer';
@@ -14,7 +14,9 @@ import LoadingCellRenderer from '../../components/LoadingCellRenderer/LoadingCel
 import PageFormat from '../../components/PageFormat/PageFormat';
 import { SideDrawerSize } from '../../components/SideDrawer/side-drawer-size';
 import SideDrawer from '../../components/SideDrawer/SideDrawer';
-import { DocumentDto, DocumentSpaceInfoDto } from '../../openapi';
+import { ToastType } from '../../components/Toast/ToastUtils/toast-type';
+import { createTextToast } from '../../components/Toast/ToastUtils/ToastUtils';
+import { DocumentDto, DocumentSpaceRequestDto, DocumentSpaceResponseDto } from '../../openapi';
 import { FormActionType } from '../../state/crud-page/form-action-type';
 import { useDocumentSpaceState } from '../../state/document-space/document-space-state';
 import { formatBytesToString } from '../../utils/file-utils';
@@ -71,7 +73,7 @@ interface DocumentSpacePageState {
   isSubmitting: boolean;
   errorMessage: string;
   showErrorMessage: boolean;
-  selectedSpace: string;
+  selectedSpace?: DocumentSpaceResponseDto;
   shouldUpdateDatasource: boolean;
   datasource?: IDatasource;
   showUploadDialog: boolean;
@@ -79,8 +81,6 @@ interface DocumentSpacePageState {
   fileToDelete: string;
   selectedFiles: DocumentDto[];
 }
-
-const selectedSpaceDefaultValue = 'Select a Space';
 
 function getDocumentUniqueKey(data: DocumentDto): string {
   return data.key;
@@ -92,7 +92,7 @@ function DocumentSpacePage() {
     isSubmitting: false,
     showErrorMessage: false,
     errorMessage: '',
-    selectedSpace: '',
+    selectedSpace: undefined,
     shouldUpdateDatasource: false,
     datasource: undefined,
     showUploadDialog: false,
@@ -103,10 +103,30 @@ function DocumentSpacePage() {
 
   const documentSpaceService = useDocumentSpaceState();
 
+  const mountedRef = useRef(false);
+
   useEffect(() => {
+    mountedRef.current = true;
+
     const spacesCancellableRequest = documentSpaceService.fetchAndStoreSpaces();
 
+    async function setInitialSpaceOnLoad() {
+      try {
+        const data = await spacesCancellableRequest.promise;
+
+        if (data && data.length > 0) {
+          setStateOnDocumentSpaceChange(data[0]);
+        }
+      } catch (err) {
+        createTextToast(ToastType.ERROR, 'Could not load Document Spaces');
+      }
+    }
+
+    setInitialSpaceOnLoad();
+
     return function cleanup() {
+      mountedRef.current = false;
+
       if (spacesCancellableRequest != null) {
         spacesCancellableRequest.cancelTokenSource.cancel();
       }
@@ -115,67 +135,80 @@ function DocumentSpacePage() {
     };
   }, []);
 
-  function onDocumentSpaceSelectionChange(
-    event: ChangeEvent<HTMLSelectElement>
-  ): void {
-    pageState.merge({
-      selectedSpace: event.target.value,
+  function mergePageState(mergeState: Partial<DocumentSpacePageState>): void {
+    if (mountedRef.current) {
+      pageState.merge(mergeState);
+    }
+  }
+
+  function setStateOnDocumentSpaceChange(documentSpace: DocumentSpaceResponseDto): void {
+    mergePageState({
+      selectedSpace: documentSpace,
       shouldUpdateDatasource: true,
       datasource: documentSpaceService.createDatasource(
-        event.target.value,
+        documentSpace.id,
         infiniteScrollOptions
       )
     });
+  }
+
+  function onDocumentSpaceSelectionChange(
+    event: ChangeEvent<HTMLSelectElement>
+  ): void {
+    const selectedDocumentSpace = documentSpaceService.documentSpaces.find(documentSpace => documentSpace.id === event.target.value);
+
+    if (selectedDocumentSpace == null) {
+      createTextToast(ToastType.ERROR, 'Could not process the selected Document Space');
+      return;
+    }
+
+    setStateOnDocumentSpaceChange(selectedDocumentSpace);
   }
 
   function onDatasourceUpdateCallback() {
     pageState.shouldUpdateDatasource.set(false);
   }
 
-  function isSelectedSpaceValid(): boolean {
-    const selectedSpace = pageState.selectedSpace.value;
-
-    return (
-      selectedSpace?.trim() !== '' &&
-      selectedSpace !== selectedSpaceDefaultValue
-    );
-  }
-
-  function getSpaceValues(): DocumentSpaceInfoDto[] {
+  function getSpaceOptions() {
     if (isDocumentSpacesLoading) {
-      return [{ name: 'Loading...' }];
+      return (
+        <option value="loading">
+          Loading...
+        </option>
+      )
     }
 
     if (isDocumentSpacesErrored) {
-      return [{ name: 'Could not load Document Spaces' }];
+      return (
+        <option value="error">
+          Could not load Document Spaces
+        </option>
+      )
     }
 
-    if (isSelectedSpaceValid()) {
-      return documentSpaceService.documentSpaces;
-    }
-
-    return [
-      { name: selectedSpaceDefaultValue },
-      ...documentSpaceService.documentSpaces,
-    ];
+    return documentSpaceService.documentSpaces.map((item) =>
+      <option key={item.id} value={item.id}>
+        {item.name}
+      </option>
+    );
   }
 
-  function submitDocumentSpace(space: DocumentSpaceInfoDto) {
+  function submitDocumentSpace(space: DocumentSpaceRequestDto) {
     pageState.merge({ isSubmitting: true });
     documentSpaceService
       .createDocumentSpace(space)
       .then((s) => {
-        pageState.merge({
+        mergePageState({
           drawerOpen: false,
           isSubmitting: false,
           showErrorMessage: false,
-          selectedSpace: s.name,
+          selectedSpace: s,
           shouldUpdateDatasource: true,
-          datasource: documentSpaceService.createDatasource(s.name, infiniteScrollOptions)
+          datasource: documentSpaceService.createDatasource(s.id, infiniteScrollOptions)
         });
       })
       .catch((message) =>
-        pageState.merge({
+        mergePageState({
           isSubmitting: false,
           errorMessage: message,
           showErrorMessage: true,
@@ -196,8 +229,14 @@ function DocumentSpacePage() {
   }
 
   async function deleteFile(): Promise<void> {
+    const selectedSpace = pageState.selectedSpace.value;
+
+    if (selectedSpace == null) {
+      return;
+    }
+
     await documentSpaceService.deleteFile(
-      pageState.selectedSpace.get(),
+      selectedSpace.id,
       pageState.fileToDelete.get()
     );
     pageState.merge({
@@ -243,17 +282,11 @@ function DocumentSpacePage() {
             <Select
               id="document-space"
               name="document-space"
-              value={pageState.selectedSpace.get()}
+              value={pageState.selectedSpace.value?.id}
               disabled={isDocumentSpacesLoading || isDocumentSpacesErrored}
               onChange={onDocumentSpaceSelectionChange}
             >
-              {getSpaceValues().map((item) => {
-                return (
-                  <option key={item.name} value={item.name}>
-                    {item.name}
-                  </option>
-                );
-              })}
+              {getSpaceOptions()}
             </Select>
             <Button
               data-testid="add-doc-space__btn"
@@ -264,9 +297,9 @@ function DocumentSpacePage() {
               Add New Space
             </Button>
           </div>
-          {isSelectedSpaceValid() &&
+          {pageState.selectedSpace.value != null &&
             <div>
-              <a href={pageState.selectedFiles.value.length > 0 ? documentSpaceService.createRelativeFilesDownloadUrl(pageState.selectedSpace.value, pageState.selectedFiles.value) : undefined}>
+              <a href={pageState.selectedFiles.value.length > 0 ? documentSpaceService.createRelativeFilesDownloadUrl(pageState.selectedSpace.value.id, pageState.selectedFiles.value) : undefined}>
                 <Button
                   data-testid="download-selected-files__btn"
                   type="button"
@@ -276,7 +309,7 @@ function DocumentSpacePage() {
                 </Button>
               </a>
 
-              <a href={documentSpaceService.createRelativeDownloadAllFilesUrl(pageState.selectedSpace.value)}>
+              <a href={documentSpaceService.createRelativeDownloadAllFilesUrl(pageState.selectedSpace.value.id)}>
                 <Button
                   data-testid="download-all-files__btn"
                   type="button"
@@ -286,7 +319,7 @@ function DocumentSpacePage() {
               </a>
 
               <DocumentUploadDialog
-                space={pageState.selectedSpace.get()}
+                documentSpaceId={pageState.selectedSpace.value.id}
                 onFinish={() => pageState.shouldUpdateDatasource.set(true)}
               />
             </div>
@@ -294,7 +327,7 @@ function DocumentSpacePage() {
         </div>
       </FormGroup>
 
-      {isSelectedSpaceValid() && pageState.datasource.value && (
+      {pageState.selectedSpace.value != null && pageState.datasource.value && (
         <InfiniteScrollGrid
           columns={documentDtoColumnsWithDelete}
           datasource={pageState.datasource.value}
